+++ python/weka/classifiers.py
@@ -23,6 +23,7 @@ import traceback
 import weka.core.jvm as jvm
 import weka.core.typeconv as typeconv
 import weka.core.classes as classes
+import weka.core.serialization as serialization
 from numpy import *
 from weka.core.classes import JavaObject, join_options, OptionHandler, Random, SelectedTag, Tags, Tag, JavaArray, \
     is_instance_of
@@ -221,6 +222,44 @@ class Classifier(OptionHandler):
             jobject=javabridge.static_call(
                 "weka/classifiers/AbstractClassifier", "makeCopy",
                 "(Lweka/classifiers/Classifier;)Lweka/classifiers/Classifier;", classifier.jobject))
+
+    @classmethod
+    def deserialize(cls, ser_file):
+        """
+        Deserializes a classifier from a file.
+
+        :param ser_file: the model file to deserialize
+        :type ser_file: str
+        :return: model and, if available, the dataset header
+        :rtype: tuple
+        """
+
+        objs = serialization.read_all(ser_file)
+        if len(objs) == 1:
+            return Classifier(jobject=objs[0]), None
+        elif len(objs) == 2:
+            return Classifier(jobject=objs[0]), Instances(jobject=objs[1])
+        else:
+            raise Exception(
+                "Excepted one or two objects in the model file (%s), but encountered: %d" % (ser_file, len(objs)))
+
+    def serialize(self, ser_file, header=None):
+        """
+        Serializes the classifier to the specified file.
+
+        :param ser_file: the file to save the model to
+        :type ser_file: str
+        :param header: the (optional) dataset header to store alongside; recommended
+        :type header: Instances
+        """
+
+        if (header is not None) and header.num_instances > 0:
+            header = Instances.template_instances(header)
+
+        if header is not None:
+            serialization.write_all(ser_file, [self, header])
+        else:
+            serialization.write(ser_file, self)
 
 
 class SingleClassifierEnhancer(Classifier):